Validation tab is now complete
