+++ opal-gwt-client/src/main/java/org/obiba/opal/web/gwt/app/client/magma/view/TableValidationView.ui.xml
@@ -1,0 +1,39 @@
+<!--
+  ~ Copyright (c) 2013 OBiBa. All rights reserved.
+  ~
+  ~ This program and the accompanying materials
+  ~ are made available under the terms of the GNU Public License v3.0.
+  ~
+  ~ You should have received a copy of the GNU General Public License
+  ~ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+  -->
+
+<ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
+    xmlns:g="urn:import:com.google.gwt.user.client.ui"
+    xmlns:b="urn:import:com.github.gwtbootstrap.client.ui"
+    ui:generateFormat='com.google.gwt.i18n.rebind.format.PropertiesFormat'
+    ui:generateKeys="com.google.gwt.i18n.server.keygen.MD5KeyGenerator"
+    ui:generateLocales="default">
+
+  <g:FlowPanel ui:field="validationPanel" styleName="min-height-450">
+      <g:FlowPanel ui:field="alertsPanel">
+          <b:Alert close="true" type="ERROR" ui:field="alert">
+              <g:Label ui:field="errorMessage"/>
+          </b:Alert>
+      </g:FlowPanel>
+
+      <g:FlowPanel addStyleNames="pull-right">
+          <b:ButtonGroup>
+              <b:Button ui:field="validate" type="INFO">
+                  <ui:msg description="Validate label">Validate</ui:msg>
+              </b:Button>
+          </b:ButtonGroup>
+      </g:FlowPanel>
+
+      <g:HTMLPanel ui:field="validationResultsPanel">
+
+      </g:HTMLPanel>
+
+  </g:FlowPanel>
+
+</ui:UiBinder>
